+++ src/app/modules/about/about-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AboutComponent } from './about/about.component';
+import { AboutComponent } from './pages/about/about.component';
 
 const routes: Routes = [{ path: '', component: AboutComponent }];
