+++ src/app/modules/about/pages/about/about.component.scss
