+++ src/app/modules/about/pages/about/about.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { MetaTagService } from 'src/app/services/meta-tag/meta-tag.service';
-import { TechIcon } from '../models/tech-icon.model';
-import { iconList } from '../data/about-website';
+import { TechIcon } from '../../models/tech-icon.model';
+import { iconList } from '../../data/about-website';
 
 @Component({
   selector: 'nm-about',
