+++ src/app/modules/projects/pages/project/project.component.ts
