+++ src/app/modules/projects/projects.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { ProjectsRoutingModule } from './projects-routing.module';
 import { ProjectsComponent } from './projects/projects.component';
 import { ProjectCardComponent } from './components/project-card/project-card.component';
-import { ProjectComponent } from './components/project/project.component';
+import { ProjectComponent } from './pages/project/project.component';
 import { NgxPageScrollModule } from 'ngx-page-scroll';
 import { NgxPageScrollCoreModule } from 'ngx-page-scroll-core';
 import { MatIconModule } from '@angular/material/icon';
