+++ src/app/modules/resume/pages/resume/resume.component.scss
