+++ src/app/modules/resume/resume-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { ResumeComponent } from './resume/resume.component';
+import { ResumeComponent } from './pages/resume/resume.component';
 
 const routes: Routes = [{ path: '', component: ResumeComponent }];
