+++ src/app/modules/resume/resume.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ResumeRoutingModule } from './resume-routing.module';
-import { ResumeComponent } from './resume/resume.component';
+import { ResumeComponent } from './pages/resume/resume.component';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatButtonModule } from '@angular/material/button';
