+++ src/app/modules/shared/shared.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AppShellNoRenderDirective } from '../../directives/app-shell/app-shell-no-render.directive';
-import { AppShellRenderDirective } from '../../directives/app-shell/app-shell-render.directive';
+import { AppShellNoRenderDirective } from './directives/app-shell/app-shell-no-render.directive';
+import { AppShellRenderDirective } from './directives/app-shell/app-shell-render.directive';
 
 @NgModule({
   declarations: [AppShellRenderDirective, AppShellNoRenderDirective],
